Move DirectVariant into the variant.rs file
